Add sorting of popular books by readers and time

+++ app/src/main/java/sg/edu/np/mad/livre/PopularBook.java
@@ -12,7 +12,7 @@ public class PopularBook {
     public String thumbnail;
     public HashMap<String, Boolean> readers;
     private int totalReaders;
-    public long totalTime;
+    public int totalTime;
 
     public int getTotalReaders() {
         return totalReaders;
